Create components for portal id and VE.Bus device
